requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# LangChain and embeddings
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-qdrant>=0.0.1
sentence-transformers>=2.2.0

# Vector database
qdrant-client>=1.7.0

# Document processing
PyMuPDF>=1.23.0
Pillow>=10.0.0
pypdf>=3.17.0

# Audio and video processing
openai-whisper>=20231117
librosa>=0.10.0
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# LLM integration
ollama>=0.1.0

# Text processing
tiktoken>=0.5.0

# Utilities
numpy>=1.24.0
python-multipart>=0.0.6

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0